requirements.txt
asgiref==3.6.0
boto3==1.26.141
botocore==1.29.141
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
crispy-bootstrap5==0.7
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==3.2.19
django-allauth==0.54.0
django-appconf==1.0.5
django-compressor==4.3.1
django-crispy-forms==2.0
django-sass-processor==1.2.2
django-storages==1.13.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
libsass==0.22.0
livereload==2.6.3
oauthlib==3.2.2
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
rcssmin==1.1.1
requests==2.30.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.5.0
urllib3==1.26.15
